refactor(showcase,core): improve context typing and update descriptions
- Replace inline empty object with a strongly-typed `EMPTY_COMPONENTS` constant to properly satisfy `CustomComponentRegistry`.
- Import `ReactNode` directly instead of accessing it via the `React` namespace in context providers.
- Populate missing `TODO` descriptions for UI components and update context file documentation.

+++ packages/core/src/engine/validators/layout-schema.ts
@@ -27,10 +27,10 @@ export const layoutSchemaValidator = z.object({
   regions: z
     .array(layoutRegionValidator)
     .min(1, 'LayoutSchema must have at least one region'),
-  gap: z.number().int().min(0).optional(),
+  gap: z.number().min(0).optional(),
   padding: z.union([
-    z.number().int().min(0),
-    z.tuple([z.number().int().min(0), z.number().int().min(0)]),
+    z.number().min(0),
+    z.tuple([z.number().min(0), z.number().min(0)]),
   ]).optional(),
   className: z.string().optional(),
   i18n: i18nConfigSchema.optional(),
